fix problem httpd.pid

+++ 80_apache_start.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
-rm /run/apache2/httpd.pid
+if [ -e /run/apache2/httpd.pid ]; then 
+  rm /run/apache2/httpd.pid
+fi
 exec /usr/sbin/httpd -D FOREGROUND -f /app/config/httpd.conf
